Code optimisation and reformat
