+++ lib/pages/history_page/bars_history_tab/bars_history_page.dart
