+++ lib/pages/wallet_page/wallet_page.dart
@@ -21,10 +21,10 @@ import '../../store/history_page_store/history_page_store.dart';
 import '../../store/market_page_store/market_page_store.dart';
 import '../../utils/card_nfc_session.dart';
 import '../../utils/header_custom_paint.dart';
+import '../../widgets/card_and_bar_tab/card_and_bar_tab.dart';
 import '../send_page/send_to/send_to_state.dart';
 import '../splash_screen/splash_screen.dart';
 import 'bar_list/bar_list.dart';
-import 'card_and_bar_tab/card_and_bar_tab.dart';
 import 'card_list/card_list.dart';
 import 'favorite_coin/favorite_coin.dart';
 import 'total_balance/total_balance.dart';
